requirements.txt
flask==3.0.0
google-genai==1.7.0
python-dotenv==1.0.0
gunicorn==21.2.0